Apply suggestions from code review

+++ src/synthpop/methods/tree_utils.py
@@ -224,6 +224,6 @@ def build_feature_matrix(X: dict[str, np.ndarray],feature_order:list[str]) -> np
     if len(set(feature_order)-set(X.keys()))>0:
         raise ValueError("cannot build feature matrix: received less columns than expected")
     if len(X.keys()) == 0:
-        return np.empty(shape=(0,0))
+        return np.empty(shape=(0,0), dtype=np.float32)
     
     return np.hstack([X[k].reshape(-1,1) if X[k].ndim==1 else X[k] for k in feature_order],dtype=np.float32)
